Reviewed by:	Chuck Murcko Submitted by:	Troy Morrison new ftp_handler.
It gets rid of the type= stuff (although it should still work -- netscape
just doesn't provide that).  It also will do an NLST if the RETR fails
(indicating that it might be a directory and not a file), and it uses a
new function (show_dir) that hacks the NLST data to provide links to the
files and directories that are returned.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@76817 13f79535-47bb-0310-9956-ffa450edef68