Reviewed by:	Randy Terbush, Brian Behlendorf Submitted by:	Dean Gaudet
Subject: warning from irix 6.2 cc

cfe: Warning 851: util.c, line 109: constant initializer expression is
invalid (refers to automatic variables).
     struct tm gmt = *gmtime(&tt);
 ----^

Interesting.  I'd have to dig out my copy of the ansi c standard to see
if this is valid -- but I'm tempted to believe the compiler because this
is a structure initialization, not a scalar.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@77465 13f79535-47bb-0310-9956-ffa450edef68