Reviewed by:	Dean Gaudet <dgaudet@arctic.org>, Jim Jagielski <jim@jaguNET.com> Submitted by:	Paul Sutton <paul@ukweb.com>
This is a repost of my patch to make httpd_monitor.c understand the
current scoreboard format (if using a scoreboard file). It also cleans up
the source a bit (removes unused defines, etc).

Also, the Makefile in the support directory still has to be edited by hand
to set the compiler name, cflags etc. Surely this can be automated by
Configure? The same information that is pre-pended to Makefile.tmpl in src
could be pre-pended to the Makefile in support.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@77550 13f79535-47bb-0310-9956-ffa450edef68