When deleting a value from a table, use table_unset() not table_set(key,NULL).
The proxy has all of its own "table" manipulation routines... like it has
struct hdr_entry, and proxy_get_header() and all this crap.  But it was
mixing the use of the real table routines and its own routines.  Clean
this up.  (Ultimately someone should clean it up to use the real table
routines, I can't see why it doesn't.)


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@79993 13f79535-47bb-0310-9956-ffa450edef68