Fix a potential memory corruption bug -- setting
    r->subprocess_env = r->main->subprocess_env

causes bogus behaviour as soon as any new values are added to
r->subprocess_env.  The new values could be allocated out of a pool
which had a shorter lifespan than the pool of the table... but only
if using the new table_xxxn API.  As a side-effect, this fix removes
the need to play with r->finfo... since modifications to the subrequest
environment don't affect the parent.

But that's a potential problem, about to post more details to new-httpd,
this is an interim fix.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@80030 13f79535-47bb-0310-9956-ffa450edef68