When writing the SERVER_SUBVERSION -D flag to ap_config.h it came out looking like:
  #define SERVER_SUBVERSION \"whatever\"

When including this in a file you end up with an unterminated string
constant.  We want it to look like this instead:

  #define SERVER_SUBVERSION "whatever"

The backslashes are only needed in the actual -D argument.

This little patch makes sure that any \" from a -D is replaced with just "
when ap_config.h is generated.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@80479 13f79535-47bb-0310-9956-ffa450edef68