So it's sunny now, and the outdoors is calling me... and any monumental fixes I've been thinking about are somewhere deep in my mind I can't find at the moment.  I'm ready to "throw in the towel" so that unix 1.3.0 can get out there, and get into reviewer's hot little hands.  I know of a few groups that want to benchmark it, but really would prefer to benchmark something released.  I really don't want to see us with win32 tied to unix release schedule forever... it's pointless.
It's all just names anyhow.  Whether we call it a beta or a current
release it's just a name.  It's time we at least had a "release candidate"
to use the 1.2 terminology... I'm with Roy, I want to drop the x.ybz
designations.  I want x.y.z, a single numbering scheme.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@80695 13f79535-47bb-0310-9956-ffa450edef68