PR: Submitted by:	Ronald.Tschalaer@psi.ch Reviewed by:	Dean Gaudet, Brian Behlendorf
Preserve the 'x-' if the client used it in its Accept-Encoding headers.
I had to make a small cosmetic change to mod_negotiation.c so that this
report would be attached to the change, since the actual one had a blank
log message.  Oops.

For more details about this whole mess, check out
<Pine.LNX.3.96dg4.980418001055.4893D-100000@twinlark.arctic.org>
From the April 1998 archives, and the whole thread around it.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@81104 13f79535-47bb-0310-9956-ffa450edef68