Add URL for "module execution order for DSO situation" patch and make sure it recognized as a showstopper because:
Without it 1.3b7 will have the same side-effects as 1.3b6 when using DSO: The
module execution order is incorrect because it is split into two separately
sorted lists and the user even has no chance to manually correct it because
the AddModule command is broken.  This patch fixes both AddModule and makes
sure the module list is reconstructed in httpd.conf when DSO's are used.

So, especially it changes some stuff in http_config a review and votes would
be fine.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@81137 13f79535-47bb-0310-9956-ffa450edef68