Ok, although no one said something about my showstopper-patch "$(LD) used but never defined" and I again had to touch more than one or two files for this (Hmmm...) I twice reviewed the patch myself now and it's correct. So, I'll commit it to remove one more release showstopper (IMHO).
The problem it addresses:

The Makefiles assumed that DSO files are build via $(LD). This is broken for
two reasons: First we never defined at least LD=ld somewhere to make sure this
works (it was silently assumed that most Make provide a built-in LD definition
- ARGL!) and second using the generic LD variable is not the truth. Instead a
special variable named LD_SHLIB is reasonable because although "ld" is
usually the default, the command for building DSO files can be "libtool" or
even "cc" on some systems.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@81270 13f79535-47bb-0310-9956-ffa450edef68