PR: Submitted by:	"Brian Havard" <brianh@kheldar.apana.org.au>
"Brian Havard" <brianh@kheldar.apana.org.au> writes:
For reasons I've yet to determine the OS/2 exe is compiled with a switch
that makes fopen() default to binary mode (-Zbin-files). This makes the !#
parser in ap_call_exec() fail on scripts that use CR/LF line terminators.
The following patch fixes that. Note that the code affected is inside an
#ifdef __EMX__ so there's no chance of breaking other platforms.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@81342 13f79535-47bb-0310-9956-ffa450edef68