Wilfredo Sanchez <wsanchez@apple.com>: ``So shared core is going to need a little work. I suggest we turn it off for now, since its easier to figure out that you can remove -s from the Makefile, than it is to make the shared core thing work out.''
So, let's do the comprimise and disable SHARED_CORE for Rhapsody/DSO situation
even when we have problems with the stripping under install time (-s).  A lot
more have to be done in 1.3.2-dev to be able to fix this. For 1.3.2 is best to
leave DSO for Rhapsody this way, i.e. without SHARED_CORE.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@82107 13f79535-47bb-0310-9956-ffa450edef68