Add more extensive checks to mod_include to avoid infinite recursive includes.
The logic used is commented in the source.  It should be complete
for the current Apache code base, but if something else starts modifying
r->filename outside of a filename translation stage that could change.

The code scans, at each stage, every subrequest for each level of
an internal redirect so it ends up being O(n^2) total in the
number of subrequests and internal redirects, but that shouldn't
be an issue.

PR: 3323
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@82369 13f79535-47bb-0310-9956-ffa450edef68