Build outside of source tree (take 3) =====================================
This patch provides a --shadow=DIR option variant for APACI in addition to the
already existing --shadow option. The difference is just this now:

   --shadow ........ creates an internal build shadow tree

   --shadow=DIR .... creates first an external build shadow tree
                     and then there the internal build shadow tree

The result is obvious: With --shadow=DIR you now can mount the extracted
Apache distribution tree read-only from NFS or CDROM or whatever and still use
the complete functionality of APACI, _INCLUDING_ creating the internal build
shadow trees to solve the parallel-platform conflict.

Just one thing we have to change in general to allow this patch: We have to
use the -h (don't follow symlinks) option of `tar' under `make install'. I
checked this against FreeBSD, Linux, SunOS and Solaris and -h really seems a
very old and well-supported option.

Now the use case:

    $ cd apache_1.3.x
    $ ./configure --shadow=/tmp/apache [...]
    $ make -f /tmp/apache/Makefile
    $ make -f /tmp/apache/Makefile install

Or alternatively:

    $ cd apache_1.3.x
    $ ./configure --shadow=/tmp/apache [...]
    $ cd /tmp/apache
    $ make
    $ make install

As you can see this is _very_ similar to the non-shadow situation ;-):

    $ cd apache_1.3.x
    $ ./configure [...]
    $ make
    $ make install


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@82449 13f79535-47bb-0310-9956-ffa450edef68