Some configure cleanups and the first attempts at some sort of universal style.
First of all, we remove the hard-wire of TAR=tar and check to see if
the tar we wind up with supports '-h'.

Secondly, when we do shell string tests, [ ".$var" != .yes ] looks better
when checking for the string "yes", I think it's better if, when checking
for/against a NULL string, we use 'x', eg:

    [ "x$var" != x ]

I'll start adjusting the other shell scripts as well.
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@82665 13f79535-47bb-0310-9956-ffa450edef68