Remove another warning:
| :> make
| gcc -c  -I../os/unix -I../include   -funsigned-char -DTARGET=\"apache\" -pipe
| -O -pedantic -Wall -Wshadow -Wpointer-arith -Wcast-align -Wmissing-prototypes
| -Wmissing-declarations -Wnested-externs -Winline `../apaci` util.c
| util.c: In function `ap_size_list_item':
| util.c:1028: warning: pointer targets in assignment differ in signedness
| util.c:1037: warning: pointer targets in assignment differ in signedness

This was caused because the involved function first does a ``const unsigned
char *ptr = (const unsigned char *)*field;'' for various reasons but later
didn't the cast again when assigning back ptr to *field. Perhaps the function
can be rewritten to not use casts at all (Ben?). But at least as long as the
first "import" cast is done we also need the two "export" casts to make the
compilers quiet.
VS: ----------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@83095 13f79535-47bb-0310-9956-ffa450edef68