OS/2: Add support for building loadable modules as DLLs.
- EXE files can't be linked to at run time so a shared core is needed.

- As uninitialized global variables can't be exported I had to add
initializers to the exported globals that didn't already have one.

- Due to an 'optimization' done for OS/2 v3, DLL file names are limited
to 8 characters plus the .dll extension. As most of the module names are
longer than that I've had to map the module names to a shorter version. This
is done in src/modules/standard/Makefile.OS2. The short version of the module
name must be used in the SharedModule statement in Configuration and in
the LoadModule directive in httpd.conf

- I had to mess with the dependencies in src/Makefile regarding the
"libhttpd.ep" as I don't want to build one. It _should_ still work for the
sysv people it's meant for but I have no way to check.

- .def files are needed for all DLLs. The one for the core dll is based on the
Win32 one and is called ApacheCoreOS2.def. The .def files for modules are
generated on the fly by the makefile.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@83178 13f79535-47bb-0310-9956-ffa450edef68