Attached is a patch for allowing user changes on the cygwin platform and a #define wrapper for the timeout singal we use to kill of pending open childs that do not react on "usual" signals.
The signalling issue seems to be a problem on the Cygwin platform, but
it's abstracted, so other platforms may benefit from it. Again,
nothing else is changed in behaviour.

Changes are:

  * src/include/ap_config.h: added the system uid for Cygwin that is
the "root" user on Cygwin

  * src/main/http_main.c: some cygwin specific #defines around
setpgrp() and getuid() calls. Adding the #define SIG_TIMEOUT_KILL to
define which singal should be used to kill of timed out childs.
Defaulting to the know value for all other plaforms.

  * src/modules/proxy/proxy_cache.c: cygwin specific #define around
setpgrp()

PR:
Obtained from:
Submitted by:	Stipe Tolj <tolj@wapme-systems.de>
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@95482 13f79535-47bb-0310-9956-ffa450edef68