Finallllly got around to testing mod_file_cache.  Looks good to me.  Both Bill's and my patches need to be merged in, but we might as well do them at the same time.  Which means somebody else needs to look at Bill's and +1 it.
Here's why it's right:  Since there's no transfer_ownership flag on
apr_mmap_dup() anymore [or rather, since it's now ignored pending an API
change at a later date to remove it], the only indication of whether an
mmap should be deleted or not is the refcount.  The core will now
delete any apr_mmap_t it encounters when it's done with it.  Before
Bill's patch, we had exactly two copies of each apr_mmap_t: one to hang
onto, and one that we passed to all requests.  But the very first
request would delete that second one, so subsequent requests would be
botched.  Each request needs its own dup of the original apr_mmap_t that
we hang onto now that there's no ownership flag to prevent a request from
deleting the shared apr_mmap_t we had before.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/APACHE_2_0_BRANCH@98530 13f79535-47bb-0310-9956-ffa450edef68