drop the guess_domain function.
Our docs say about AuthDigestDomain:
This directive should always be specified and contain at least the (set of)
root URI(s) for this space. Omitting to do so will cause the client to send
the Authorization header for every request sent to this server.

guessing the parameter is somewhat bogus. guess_domain() also resulted sometimes
in relative URIs, non-URI strings or empty strings, which caused a lot of
problems.
According to the docs, the domain parameter will be omitted now,
if not specified. This is exactly, what one would expect.

PR: 16937 (related to)


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@98677 13f79535-47bb-0310-9956-ffa450edef68