I'm almost done here.
Fixup the vhost example.

Fix comments in the main config.  In particular, move the User/Group
directive out of the "You can override this in a vhost" section.
(That was probably a legacy of the days when User/Group was used
inside a vhost for suexec.)  Remove the "Action" examples,
which didn't really provide enough information to make any use of
those directives.  And overall cleanup.


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/simple-conf@160210 13f79535-47bb-0310-9956-ffa450edef68