Merge r1429582 from trunk:
According to my testing, one special case of 'log_xlate_error', i.e. EES_INCOMPLETE_CHAR,
 is 13x (!!!) faster with the use 'ap_bin2hex' instead of apr_snprintf(..., "%02X" + srlen for each character.

Output is *not* exactly the same. It was uppercase, now it is lowercase.
It is just for logging, so I don't think it is an issue.
Should it be, a call to ap_strtoupper can be added.

So sad it is just for logging in case of error... no real speedup to be expected in real life .

Submitted by: jailletc36
Reviewed/backported by: jim


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1455221 13f79535-47bb-0310-9956-ffa450edef68