Merge r1491221, r1491479, r1491538 from trunk:
Reduce stack usage, 32 bytes are enough here, no need to take 8k in the stack. (19 would be enough, but be safe...)

Use the size of the buffer as opposed to MAX_STRING_LEN.


Reduce stack usage, 32 bytes are enough here, no need to
take 8k in the stack. (19 would be enough, but be safe...)
Use the size of the buffer as opposed to MAX_STRING_LEN.

Followup to r1491221+r1491479: change consistently
the same code used in another place in the same file.

Submitted by: jailletc36, minfrin, rjung
Reviewed/backported by: jim


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1492198 13f79535-47bb-0310-9956-ffa450edef68